Bot/requirements.txt
aiogram==2.22.2
aiohttp==3.8.3
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
Babel==2.9.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.5
docopt==0.6.2
frozenlist==1.3.1
idna==3.4
multidict==6.0.2
num2words==0.5.12
numpy==1.23.3
pycparser==2.21
python-dotenv==0.21.0
pytz==2022.4
requests==2.28.1
srt==3.5.2
torch==1.12.1
tqdm==4.64.1
typing_extensions==4.3.0
urllib3==1.26.12
vosk==0.3.44
websockets==10.3
yarl==1.8.1
openai
